webapp/requirements.txt
altair==5.3.0
attrs==23.2.0
beautifulsoup4==4.12.3
black==24.3.0
blinker==1.7.0
cachetools==5.3.3
captcha==0.5.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dnspython==2.6.1
email_validator==2.1.1
entrypoints==0.4
Faker==24.4.0
favicon==0.7.0
fonttools==4.50.0
gitdb==4.0.11
GitPython==3.1.43
htbuilder==0.6.2
idna==3.6
Jinja2==3.1.3
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lxml==5.2.1
Markdown==3.6
markdown-it-py==3.0.0
markdownlit==0.0.7
MarkupSafe==2.1.5
matplotlib==3.8.3
mdurl==0.1.2
more-itertools==10.2.0
mypy==1.9.0
mypy-extensions==1.0.0
numpy==1.26.4
packaging==23.2
pandas==2.2.1
pandas-stubs==2.2.1.240316
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
plotly==5.20.0
prometheus_client==0.20.0
protobuf==4.25.3
pyarrow==15.0.2
pycountry==23.12.11
pydeck==0.8.1b0
Pygments==2.17.2
pymdown-extensions==10.7.1
pymongo==4.6.3
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
referencing==0.34.0
requests==2.31.0
rich==13.7.1
rpds-py==0.18.0
setuptools==68.2.2
six==1.16.0
smmap==5.0.1
soupsieve==2.5
st-annotated-text==4.0.1
streamlit==1.32.2
streamlit-camera-input-live==0.2.0
streamlit-card==1.0.0
streamlit-embedcode==0.1.2
streamlit-extras==0.4.0
streamlit-faker==0.0.3
streamlit-image-coordinates==0.1.6
streamlit-keyup==0.2.4
streamlit-toggle-switch==1.0.2
streamlit-vertical-slider==2.5.5
tenacity==8.2.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.1
tornado==6.4
types-pytz==2024.1.0.20240203
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1